App component moved to an individual folder
