+++ src/main.ts
@@ -1,8 +1,8 @@
 import { mount } from "svelte";
-import App from "./App.svelte";
+import App from "./App";
 
 const app = mount(App, {
-  target: document.getElementById("app")!,
+	target: document.getElementById("app")!,
 });
 
 export default app;
